+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### WIP
 - Clean npm cache for `npm:clean`
 - Add structure to `server` directory
+- Add Prerender support (server and middleware)
 
 #### 0.6.6 (January 31, 2014)
 - Remove Autoprefixer for now